Problem: spi_executors result don't enforce return types
It seem to have done this accidentally in some tests, but simply
observing that we're iterating over `Args` to get types shows us that
this is wrong.

Solution: ensure we actually iterate the return tuple types instead

This necessitated some signature changes (dropping `Args` is obvious)
and promotion of `cppgres::oid` into a struct so we can implement
correct conversion and type traits on it.

+++ tests/xact.hpp
@@ -35,7 +35,7 @@ add_test(internal_subtransaction_commit, ([](test_case &) {
              spi.execute("insert into internal_subtransaction_commit default values");
            }
            cppgres::spi_executor spi;
-           auto res = spi.query<std::tuple<int32_t>>(
+           auto res = spi.query<std::tuple<int64_t>>(
                "select count(*) from internal_subtransaction_commit");
            result = result && _assert(std::get<0>(res.begin()[0]) == 1);
            return result;
@@ -53,7 +53,7 @@ add_test(internal_subtransaction_rollback, ([](test_case &) {
              spi.execute("insert into internal_subtransaction_rollback default values");
            }
            cppgres::spi_executor spi;
-           auto res = spi.query<std::tuple<int32_t>>(
+           auto res = spi.query<std::tuple<int64_t>>(
                "select count(*) from internal_subtransaction_rollback");
            result = result && _assert(std::get<0>(res.begin()[0]) == 0);
            return result;
